Address audit review findings
Tighten workflow permissions, bound external reads and pagination, and fix date, authentication, and record-cleanup edge cases identified during review.

+++ apps/web/src/lib/search-results.ts
@@ -14,7 +14,7 @@ import { z } from "zod";
 export const searchRouteSchema = z.object({
 	q: z.string().optional(),
 	type: z.string().optional(),
-	page: z.coerce.number().min(1).optional().default(1),
+	page: z.coerce.number().int().min(1).optional().default(1),
 });
 
 export type SearchRouteParams = z.infer<typeof searchRouteSchema>;
